chore: change folder names
We defined the app namespace, which implies to change the directories name. So does this commit
