+++ app/src/test/java/ch/hikemate/app/screen/SecondScreen.kt
@@ -1,7 +1,7 @@
-package com.android.sample.screen
+package ch.hikemate.app.screen
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
-import com.android.sample.resources.C
+import ch.hikemate.app.resources.C
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
